add precision support

+++ src/integration-test/kotlin/integration/btc/BtcNotaryIntegrationTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.fail
 import util.getRandomString
+import java.math.BigDecimal
 import java.math.BigInteger
 
 
@@ -34,7 +35,7 @@ class BtcNotaryIntegrationTest {
         Thread.sleep(20_000)
         val newBalance = integrationHelper.getIrohaAccountBalance(testClient, btcAsset)
         assertEquals(
-            initialBalance.add(BigInteger.valueOf(btcToSat(btcAmount))),
+            BigDecimal(initialBalance).add(BigDecimal.valueOf(btcToSat(btcAmount))),
             newBalance
         )
     }
